Extract SelfRegistering class
